api/requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra=dev --extra=test --output-file=requirements/dev.txt pyproject.toml
#
aiofiles==23.1.0
    # via verdantech-api (pyproject.toml)
aiosmtplib==2.0.2
    # via verdantech-api (pyproject.toml)
alembic==1.11.2
    # via litestar
anyio==3.7.1
    # via
    #   httpcore
    #   litestar
    #   watchfiles
argon2-cffi==21.3.0
    # via verdantech-api (pyproject.toml)
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
black==23.7.0
    # via verdantech-api (pyproject.toml)
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via argon2-cffi-bindings
charset-normalizer==3.2.0
    # via requests
click==8.1.6
    # via
    #   black
    #   litestar
    #   mkdocs
    #   rich-click
    #   uvicorn
colorama==0.4.6
    # via
    #   griffe
    #   mkdocs-material
coverage[toml]==7.2.7
    # via verdantech-api (pyproject.toml)
dnspython==2.4.1
    # via email-validator
editorconfig==0.12.3
    # via jsbeautifier
email-validator==2.0.0.post2
    # via verdantech-api (pyproject.toml)
exceptiongroup==1.1.2
    # via
    #   anyio
    #   pytest
factory-boy==3.3.0
    # via verdantech-api (pyproject.toml)
faker==19.2.0
    # via
    #   factory-boy
    #   polyfactory
fast-query-parsers==1.0.2
    # via litestar
fastjsonschema==2.18.0
    # via validate-pyproject
ghp-import==2.1.0
    # via mkdocs
greenlet==2.0.2
    # via sqlalchemy
griffe==0.32.3
    # via mkdocstrings-python
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
html2text==2020.1.16
    # via verdantech-api (pyproject.toml)
httpcore==0.17.3
    # via httpx
httptools==0.6.0
    # via uvicorn
httpx==0.24.1
    # via litestar
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via verdantech-api (pyproject.toml)
jinja2==3.1.2
    # via
    #   mkdocs
    #   mkdocs-material
    #   mkdocstrings
jsbeautifier==1.14.9
    # via litestar
litestar[cli,sqlalchemy]==2.0.0rc1
    # via verdantech-api (pyproject.toml)
mako==1.2.4
    # via alembic
markdown==3.4.4
    # via
    #   mkdocs
    #   mkdocs-autorefs
    #   mkdocs-material
    #   mkdocstrings
    #   pymdown-extensions
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via
    #   jinja2
    #   mako
    #   mkdocs
    #   mkdocstrings
mdurl==0.1.2
    # via markdown-it-py
mergedeep==1.3.4
    # via mkdocs
mkdocs==1.5.2
    # via
    #   mkdocs-autorefs
    #   mkdocs-material
    #   mkdocstrings
    #   verdantech-api (pyproject.toml)
mkdocs-autorefs==0.5.0
    # via mkdocstrings
mkdocs-material==9.1.21
    # via verdantech-api (pyproject.toml)
mkdocs-material-extensions==1.1.1
    # via mkdocs-material
mkdocstrings[python]==0.22.0
    # via
    #   mkdocstrings-python
    #   verdantech-api (pyproject.toml)
mkdocstrings-python==1.2.1
    # via mkdocstrings
msgspec==0.17.0
    # via litestar
multidict==6.0.4
    # via litestar
mypy-extensions==1.0.0
    # via
    #   black
    #   typing-inspect
packaging==23.1
    # via
    #   black
    #   mkdocs
    #   pytest
passlib==1.7.4
    # via verdantech-api (pyproject.toml)
pathspec==0.11.2
    # via
    #   black
    #   mkdocs
platformdirs==3.10.0
    # via
    #   black
    #   mkdocs
pluggy==1.2.0
    # via pytest
polyfactory==2.7.0
    # via litestar
py==1.11.0
    # via pytest-html
pycparser==2.21
    # via cffi
pygments==2.15.1
    # via
    #   mkdocs-material
    #   rich
pymdown-extensions==10.1
    # via
    #   mkdocs-material
    #   mkdocstrings
pytest==7.4.0
    # via
    #   pytest-asyncio
    #   pytest-html
    #   pytest-metadata
    #   pytest-mock
    #   verdantech-api (pyproject.toml)
pytest-asyncio==0.21.1
    # via verdantech-api (pyproject.toml)
pytest-html==3.2.0
    # via verdantech-api (pyproject.toml)
pytest-metadata==3.0.0
    # via pytest-html
pytest-mock==3.11.1
    # via verdantech-api (pyproject.toml)
python-dateutil==2.8.2
    # via
    #   faker
    #   ghp-import
    #   serpyco
python-decouple==3.8
    # via verdantech-api (pyproject.toml)
python-dotenv==1.0.0
    # via uvicorn
python-rapidjson==1.10
    # via serpyco
pyyaml==6.0.1
    # via
    #   litestar
    #   mkdocs
    #   pymdown-extensions
    #   pyyaml-env-tag
    #   uvicorn
pyyaml-env-tag==0.1
    # via mkdocs
regex==2023.6.3
    # via mkdocs-material
requests==2.31.0
    # via mkdocs-material
rich==13.5.2
    # via
    #   litestar
    #   rich-click
rich-click==1.6.1
    # via litestar
ruff==0.0.282
    # via verdantech-api (pyproject.toml)
serpyco==1.3.10
    # via verdantech-api (pyproject.toml)
six==1.16.0
    # via
    #   jsbeautifier
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==2.0.19
    # via
    #   alembic
    #   litestar
structlog==23.1.0
    # via verdantech-api (pyproject.toml)
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   pytest
typing-extensions==4.7.1
    # via
    #   alembic
    #   litestar
    #   polyfactory
    #   sqlalchemy
    #   typing-inspect
    #   uvicorn
typing-inspect==0.9.0
    # via serpyco
urllib3==2.0.4
    # via requests
uvicorn[standard]==0.23.2
    # via litestar
uvloop==0.17.0
    # via uvicorn
validate-pyproject==0.13
    # via verdantech-api (pyproject.toml)
watchdog==3.0.0
    # via mkdocs
watchfiles==0.19.0
    # via uvicorn
websockets==11.0.3
    # via uvicorn
